pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "django-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"django>=5.1.6",
	"djangorestframework-simplejwt>=5.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]  # Tells setuptools to look inside "src" for packages

[dependency-groups]
dev = [
	"mypy>=1.15.0",
	"ruff>=0.9.9",
	"invoke>=2.2.0",
	"pytest>=8.3.5",
	"pytest-django>=4.10.0",
	"faker>=36.2.2",
	"freezegun>=1.5.1",
	"django-stubs[compatible-mypy]>=5.1.3",
]

[tool.ruff]
preview = true
line-length = 88
exclude = [
	".git",
	".venv",
	"**/migrations/",
	"**project/settings.py",
]
target-version = "py312"

[tool.ruff.lint]
select = [
	"I",
	"F",
	"E",
	"W",
	"Q",
	"UP",
	"FAST",
	"LOG",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = [
	"tests",
]

[tool.django-stubs]
django_settings_module = "project.settings"

[tool.mypy]
exclude = [
	".venv/",
	"migrations/",
	"models/",
]
ignore_missing_imports = true
allow_subclassing_any = false

[mypy.plugins.django-stubs]
django_settings_module = "project.settings"
